requirements.txt
customtkinter
mysql-connector-python
pymongo
bcrypt
matplotlib
